+++ SpiNNaker-machine/src/main/java/uk/ac/manchester/spinnaker/machine/temp/package-info.java
@@ -1,0 +1,2 @@
+/** Temporary Stuff which will be DELETED!. */
+package uk.ac.manchester.spinnaker.machine.temp;
